add option for rpi

+++ utility/xubuntu_armbian_install.sh
@@ -35,11 +35,14 @@ sudo apt-get -y install gnome-bluetooth indicator-bluetooth
 #install update manger 
 sudo apt-get -y install ubuntu-advantage-tools
 
-# remove last cloud-init and run netplan, desactivate networkd
-sudo rm /etc/netplan/50-cloud-init.yaml
-sudo netplan generate
-sudo systemctl stop systemd-networkd-wait-online.service
-sudo systemctl disable systemd-networkd-wait-online.service
-sudo rm /etc/network/interfaces.d/99-add-config.conf
-sudo echo "# The loopback network interface\nauto lo\niface lo inet loopback\n\n# The primary network interface\nauto eth0\niface eth0 inet dhcp\n" > /tmp/99-add-config.conf
-sudo mv /tmp/99-add-config.conf /etc/network/interfaces.d/99-add-config.conf
+if ["$1" = "rpi"]
+then
+  # remove last cloud-init and run netplan, desactivate networkd
+  sudo rm /etc/netplan/50-cloud-init.yaml
+  sudo netplan generate
+  sudo systemctl stop systemd-networkd-wait-online.service
+  sudo systemctl disable systemd-networkd-wait-online.service
+  sudo rm /etc/network/interfaces.d/99-add-config.conf
+  sudo echo "# The loopback network interface\nauto lo\niface lo inet loopback\n\n# The primary network interface\nauto eth0\niface eth0 inet dhcp\n" > /tmp/99-add-config.conf
+  sudo mv /tmp/99-add-config.conf /etc/network/interfaces.d/99-add-config.conf
+fi
